+++ event-emitter-sample/src/main/java/com/zhuinden/eventemittersample/core/navigation/BaseFragment.kt
@@ -8,5 +8,6 @@ import com.zhuinden.eventemittersample.utils.requireArguments
  */
 
 open class BaseFragment : Fragment() {
+    @Suppress("NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS")
     fun <T : BaseKey> getKey(): T = requireArguments.getParcelable<T>("KEY")
 }