+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+-Event Emitter 1.0.0 (2019-09-27)
+--------------------------------
+- Update command-queue to 1.0.0 (which was just a version bump).
+
+- Revise `event-emitter-sample` to use `LifecycleObserver` and `LifecycleOwner` rather than `CompositeNotificationToken`.
+
+
 -Event Emitter 0.0.1 (2019-06-04)
 --------------------------------
-- Initial Release.
+- Initial release.